Refactor task publish rucio (#7849)
* beware missing nextIterFiles in summary

* introduce PublisherUtils.py and use in TaskPublishRucio.py

* introduce PublisherDbsUtils.py and use in TaskPublishRucio.py

* cleanup blockname vs block_name

* fix indent bug in TaskPublishRucio.py
